+++ .aegir.js
@@ -1,15 +1,16 @@
 'use strict'
 
-const PeerInfo = require('peer-info')
-const PeerId = require('peer-id')
 const pull = require('pull-stream')
 const parallel = require('async/parallel')
-
 const WebSocketStarRendezvous = require('libp2p-websocket-star-rendezvous')
 const sigServer = require('libp2p-webrtc-star/src/sig-server')
 
-const rawPeer = require('./test/fixtures/test-peer.json')
 const Node = require('./test/utils/bundle-nodejs.js')
+const {
+  getPeerRelay,
+  WRTC_RENDEZVOUS_MULTIADDR,
+  WS_RENDEZVOUS_MULTIADDR
+} = require('./test/utils/constants')
 
 let wrtcRendezvous
 let wsRendezvous
@@ -19,7 +20,7 @@ const before = (done) => {
   parallel([
     (cb) => {
       sigServer.start({
-        port: 15555
+        port: WRTC_RENDEZVOUS_MULTIADDR.nodeAddress().port
         // cryptoChallenge: true TODO: needs https://github.com/libp2p/js-libp2p-webrtc-star/issues/128
       }, (err, server) => {
         if (err) {
@@ -31,7 +32,7 @@ const before = (done) => {
     },
     (cb) => {
       WebSocketStarRendezvous.start({
-        port: 14444,
+        port: WS_RENDEZVOUS_MULTIADDR.nodeAddress().port,
         refreshPeerListIntervalMS: 1000,
         strictMultiaddr: false,
         cryptoChallenge: true
@@ -44,17 +45,24 @@ const before = (done) => {
       })
     },
     (cb) => {
-      PeerId.createFromJSON(rawPeer, (err, peerId) => {
+      getPeerRelay((err, peerInfo) => {
         if (err) {
           return done(err)
         }
-        const peer = new PeerInfo(peerId)
-
-        peer.multiaddrs.add('/ip4/127.0.0.1/tcp/9200/ws')
 
         node = new Node({
-          peerInfo: peer
+          peerInfo,
+          config: {
+            relay: {
+              enabled: true,
+              hop: {
+                enabled: true,
+                active: true
+              }
+            }
+          }
         })
+
         node.handle('/echo/1.0.0', (protocol, conn) => pull(conn, conn))
         node.start(cb)
       })
@@ -71,7 +79,7 @@ const after = (done) => {
 }
 
 module.exports = {
-  bundlesize: { maxSize: '215kB' },
+  bundlesize: { maxSize: '218kB' },
   hooks: {
     pre: before,
     post: after