+++ src/config.js
@@ -1,77 +1,84 @@
 'use strict'
 
 const { struct, superstruct } = require('superstruct')
-const kind = require('kind-of')
 const { optional, list } = struct
 
-const transports = ['tcp', 'utp', 'webrtcstar', 'webrtcdirect', 'websockets', 'websocketstar']
 // Define custom types
-const s = superstruct({
-  types: {
-    tcp: v => kind(v) === 'tcp',
-    utp: v => kind(v) === 'utp',
-    webrtcstar: v => kind(v) === 'webrtcstar',
-    webrtcdirect: v => kind(v) === 'webrtcdirect',
-    websockets: v => kind(v) === 'websockets',
-    websocketstar: v => kind(v) === 'websocketstar',
-    transport: value => {
-      const [error] = list([s.union([ ...transports, 'function' ])]).validate(value)
-      if (error) return error.message
-
-      return value.length > 0
-        ? true
-        : 'You need to provide at least one transport.'
+const s = superstruct()
+const transport = s.union([
+  s.interface({
+    createListener: 'function',
+    dial: 'function'
+  }),
+  'function'
+])
+const modulesSchema = s({
+  connEncryption: optional(list([s('object|function')])),
+  // this is hacky to simulate optional because interface doesnt work correctly with it
+  // change to optional when fixed upstream
+  connProtector: s.union(['undefined', s.interface({ protect: 'function' })]),
+  contentRouting: optional(list(['object'])),
+  dht: optional(s('null|function|object')),
+  peerDiscovery: optional(list([s('object|function')])),
+  peerRouting: optional(list(['object'])),
+  streamMuxer: optional(list([s('object|function')])),
+  transport: s.intersection([[transport], s.interface({
+    length (v) {
+      return v > 0 ? true : 'ERROR_EMPTY'
     }
-  }
+  })])
 })
 
-const optionsSchema = s(
-  {
-    connectionManager: 'object?',
-    datastore: 'object?',
-    peerInfo: 'object',
-    peerBook: 'object?',
-    modules: s({
-      connEncryption: optional(list([s('object|function')])),
-      // this is hacky to simulate optional because interface doesnt work correctly with it
-      // change to optional when fixed upstream
-      connProtector: s.union(['undefined', s.interface({ protect: 'function' })]),
-      contentRouting: optional(list(['object'])),
-      dht: optional(s('null|function|object')),
-      peerDiscovery: optional(list([s('object|function')])),
-      peerRouting: optional(list(['object'])),
-      streamMuxer: optional(list([s('object|function')])),
-      transport: 'transport'
-    }),
-    config: s({
-      peerDiscovery: 'object?',
-      relay: s({
-        enabled: 'boolean',
-        hop: optional(s({
-          enabled: 'boolean',
-          active: 'boolean'
-        },
-        { enabled: false, active: false }))
-      }, { enabled: true, hop: {} }),
-      dht: s({
-        kBucketSize: 'number',
-        enabled: 'boolean?',
-        randomWalk: optional(s({
-          enabled: 'boolean?',
-          queriesPerPeriod: 'number?',
-          interval: 'number?',
-          timeout: 'number?'
-        }, { enabled: true, queriesPerPeriod: 1, interval: 30000, timeout: 10000 })),
-        validators: 'object?',
-        selectors: 'object?'
-      }, { enabled: true, kBucketSize: 20, enabledDiscovery: true }),
-      EXPERIMENTAL: s({
-        pubsub: 'boolean'
-      }, { pubsub: false })
-    }, { relay: {}, dht: {}, EXPERIMENTAL: {} })
-  },
-  { config: {}, modules: {} }
-)
+const configSchema = s({
+  peerDiscovery: s('object', {
+    autoDial: true
+  }),
+  relay: s({
+    enabled: 'boolean',
+    hop: optional(s({
+      enabled: 'boolean',
+      active: 'boolean'
+    }, {
+      // HOP defaults
+      enabled: false,
+      active: false
+    }))
+  }, {
+    // Relay defaults
+    enabled: true
+  }),
+  // DHT config
+  dht: s('object?', {
+    // DHT defaults
+    enabled: false,
+    kBucketSize: 20,
+    randomWalk: {
+      enabled: false, // disabled waiting for https://github.com/libp2p/js-libp2p-kad-dht/issues/86
+      queriesPerPeriod: 1,
+      interval: 300e3,
+      timeout: 10e3
+    }
+  }),
+  // Experimental config
+  EXPERIMENTAL: s({
+    pubsub: 'boolean'
+  }, {
+    // Experimental defaults
+    pubsub: false
+  })
+}, {})
+
+const optionsSchema = s({
+  switch: 'object?',
+  connectionManager: s('object', {
+    minPeers: 25
+  }),
+  datastore: 'object?',
+  peerInfo: 'object',
+  peerBook: 'object?',
+  modules: modulesSchema,
+  config: configSchema
+})
 
 module.exports.validate = (opts) => {
   const [error, options] = optionsSchema.validate(opts)
@@ -84,6 +91,10 @@ module.exports.validate = (opts) => {
     if (options.config.dht.enabled) {
       s('function|object')(options.modules.dht)
     }
+  }
+
+  if (options.config.peerDiscovery.autoDial === undefined) {
+    options.config.peerDiscovery.autoDial = true
   }
 
   return options