+++ src/dht.js
@@ -1,10 +1,15 @@
 'use strict'
 
+const nextTick = require('async/nextTick')
+const errCode = require('err-code')
+
+const { messages, codes } = require('./errors')
+
 module.exports = (node) => {
   return {
     put: (key, value, callback) => {
       if (!node._dht) {
-        return callback(new Error('DHT is not available'))
+        return nextTick(callback, errCode(new Error(messages.DHT_DISABLED), codes.DHT_DISABLED))
       }
 
       node._dht.put(key, value, callback)
@@ -16,7 +21,7 @@ module.exports = (node) => {
       }
 
       if (!node._dht) {
-        return callback(new Error('DHT is not available'))
+        return nextTick(callback, errCode(new Error(messages.DHT_DISABLED), codes.DHT_DISABLED))
       }
 
       node._dht.get(key, options, callback)
@@ -28,7 +33,7 @@ module.exports = (node) => {
       }
 
       if (!node._dht) {
-        return callback(new Error('DHT is not available'))
+        return nextTick(callback, errCode(new Error(messages.DHT_DISABLED), codes.DHT_DISABLED))
       }
 
       node._dht.getMany(key, nVals, options, callback)