+++ src/errors.js
@@ -1,0 +1,13 @@
+'use strict'
+
+exports.messages = {
+  NOT_STARTED_YET: 'The libp2p node is not started yet',
+  DHT_DISABLED: 'DHT is not available'
+}
+
+exports.codes = {
+  DHT_DISABLED: 'ERR_DHT_DISABLED',
+  PUBSUB_NOT_STARTED: 'ERR_PUBSUB_NOT_STARTED',
+  ERR_NODE_NOT_STARTED: 'ERR_NODE_NOT_STARTED',
+  ERR_DISCOVERED_SELF: 'ERR_DISCOVERED_SELF'
+}