+++ test/browser.js
@@ -1,3 +1,4 @@
 'use strict'
 
+require('./circuit-relay.browser')
 require('./transports.browser')