+++ test/peer-discovery.node.js
@@ -9,15 +9,13 @@ const signalling = require('libp2p-webrtc-star/src/sig-server')
 const parallel = require('async/parallel')
 const crypto = require('crypto')
 
-const PeerId = require('peer-id')
-const PeerInfo = require('peer-info')
-
 const createNode = require('./utils/create-node')
 const echo = require('./utils/echo')
 
 describe('peer discovery', () => {
   let nodeA
   let nodeB
+  let nodeC
   let port = 24642
   let ss
 
@@ -49,6 +47,15 @@ describe('peer discovery', () => {
           nodeB = node
           node.handle('/echo/1.0.0', echo)
           node.start(cb)
+        }),
+        (cb) => createNode([
+          '/ip4/0.0.0.0/tcp/0',
+          `/ip4/127.0.0.1/tcp/${port}/ws/p2p-webrtc-star`
+        ], options, (err, node) => {
+          expect(err).to.not.exist()
+          nodeC = node
+          node.handle('/echo/1.0.0', echo)
+          node.start(cb)
         })
       ], done)
     })
@@ -57,8 +64,13 @@ describe('peer discovery', () => {
       parallel([
         (cb) => nodeA.stop(cb),
         (cb) => nodeB.stop(cb),
+        (cb) => nodeC.stop(cb),
         (cb) => ss.stop(cb)
       ], done)
+    })
+
+    afterEach(() => {
+      sinon.restore()
     })
   }
 
@@ -66,6 +78,7 @@ describe('peer discovery', () => {
     it('should enable by default a module passed as an object', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0)
       }
@@ -86,6 +99,7 @@ describe('peer discovery', () => {
     it('should enable by default a module passed as a function', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0)
       }
@@ -108,6 +122,7 @@ describe('peer discovery', () => {
     it('should enable module by configutation', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0),
         tag: 'mockDiscovery'
@@ -143,6 +158,7 @@ describe('peer discovery', () => {
     it('should disable module by configutation', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0),
         tag: 'mockDiscovery'
@@ -178,6 +194,7 @@ describe('peer discovery', () => {
     it('should register module passed as function', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0)
       }
@@ -215,6 +232,7 @@ describe('peer discovery', () => {
     it('should register module passed as object', (done) => {
       const mockDiscovery = {
         on: sinon.stub(),
+        removeListener: sinon.stub(),
         start: sinon.stub().callsArg(0),
         stop: sinon.stub().callsArg(0),
         tag: 'mockDiscovery'
@@ -241,6 +259,33 @@ describe('peer discovery', () => {
     })
   })
 
+  describe('discovery scenarios', () => {
+    setup({
+      config: {
+        dht: {
+          enabled: false
+        },
+        peerDiscovery: {
+          autoDial: false,
+          bootstrap: {
+            enabled: true,
+            list: []
+          }
+        }
+      }
+    })
+
+    it('should ignore self on discovery', function () {
+      const discoverySpy = sinon.spy()
+      nodeA.on('peer:discovery', discoverySpy)
+      nodeA._discovery[0].emit('peer', nodeA.peerInfo)
+
+      expect(discoverySpy.called).to.eql(false)
+      expect(nodeA.peerBook.getAllArray()).to.have.length(0)
+      expect()
+    })
+  })
+
   describe('MulticastDNS', () => {
     setup({
       config: {
@@ -248,8 +293,10 @@ describe('peer discovery', () => {
           enabled: false
         },
         peerDiscovery: {
+          autoDial: true,
           mdns: {
             enabled: true,
+            interval: 200, // discover quickly
             // use a random tag to prevent CI collision
             serviceTag: crypto.randomBytes(10).toString('hex')
           }
@@ -257,13 +304,23 @@ describe('peer discovery', () => {
       }
     })
 
-    it('find a peer', function (done) {
-      this.timeout(15 * 1000)
+    it('find peers', function (done) {
+      let expectedPeers = new Set([
+        nodeB.peerInfo.id.toB58String(),
+        nodeC.peerInfo.id.toB58String()
+      ])
 
-      nodeA.once('peer:discovery', (peerInfo) => {
-        expect(nodeB.peerInfo.id.toB58String())
-          .to.eql(peerInfo.id.toB58String())
+      function finish () {
+        nodeA.removeAllListeners('peer:discovery')
+        expect(expectedPeers.size).to.eql(0)
         done()
+      }
+
+      nodeA.on('peer:discovery', (peerInfo) => {
+        expectedPeers.delete(peerInfo.id.toB58String())
+        if (expectedPeers.size === 0) {
+          finish()
+        }
       })
     })
   })
@@ -276,6 +333,7 @@ describe('peer discovery', () => {
           enabled: false
         },
         peerDiscovery: {
+          autoDial: true,
           webRTCStar: {
             enabled: true
           }
@@ -283,12 +341,24 @@ describe('peer discovery', () => {
       }
     })
 
-    it('find a peer', function (done) {
-      this.timeout(15 * 1000)
-      nodeA.once('peer:discovery', (peerInfo) => {
-        expect(nodeB.peerInfo.id.toB58String())
-          .to.eql(peerInfo.id.toB58String())
+    it('find peers', function (done) {
+      this.timeout(20e3)
+      let expectedPeers = new Set([
+        nodeB.peerInfo.id.toB58String(),
+        nodeC.peerInfo.id.toB58String()
+      ])
+
+      function finish () {
+        nodeA.removeAllListeners('peer:discovery')
+        expect(expectedPeers.size).to.eql(0)
         done()
+      }
+
+      nodeA.on('peer:discovery', (peerInfo) => {
+        expectedPeers.delete(peerInfo.id.toB58String())
+        if (expectedPeers.size === 0) {
+          finish()
+        }
       })
     })
   })
@@ -300,8 +370,10 @@ describe('peer discovery', () => {
           enabled: false
         },
         peerDiscovery: {
+          autoDial: true,
           mdns: {
             enabled: true,
+            interval: 200, // discovery quickly
             // use a random tag to prevent CI collision
             serviceTag: crypto.randomBytes(10).toString('hex')
           },
@@ -312,12 +384,23 @@ describe('peer discovery', () => {
       }
     })
 
-    it('find a peer', function (done) {
-      this.timeout(15 * 1000)
-      nodeA.once('peer:discovery', (peerInfo) => {
-        expect(nodeB.peerInfo.id.toB58String())
-          .to.eql(peerInfo.id.toB58String())
+    it('find peers', function (done) {
+      let expectedPeers = new Set([
+        nodeB.peerInfo.id.toB58String(),
+        nodeC.peerInfo.id.toB58String()
+      ])
+
+      function finish () {
+        nodeA.removeAllListeners('peer:discovery')
+        expect(expectedPeers.size).to.eql(0)
         done()
+      }
+
+      nodeA.on('peer:discovery', (peerInfo) => {
+        expectedPeers.delete(peerInfo.id.toB58String())
+        if (expectedPeers.size === 0) {
+          finish()
+        }
       })
     })
   })
@@ -326,32 +409,97 @@ describe('peer discovery', () => {
     setup({
       config: {
         peerDiscovery: {
+          autoDial: true,
           mdns: {
             enabled: false
           },
           webRTCStar: {
             enabled: false
           }
+        },
+        dht: {
+          enabled: true,
+          kBucketSize: 20,
+          randomWalk: {
+            enabled: true,
+            queriesPerPeriod: 1,
+            delay: 100,
+            interval: 200, // start the query sooner
+            timeout: 3000
+          }
         }
       }
     })
 
-    it('find a peer', function (done) {
-      this.timeout(15 * 1000)
+    it('find peers through the dht', function (done) {
+      let expectedPeers = new Set([
+        nodeB.peerInfo.id.toB58String(),
+        nodeC.peerInfo.id.toB58String()
+      ])
 
-      nodeA.once('peer:discovery', (peerInfo) => {
-        expect(nodeB.peerInfo.id.toB58String())
-          .to.eql(peerInfo.id.toB58String())
+      function finish () {
+        nodeA.removeAllListeners('peer:discovery')
+        expect(expectedPeers.size).to.eql(0)
         done()
+      }
+
+      nodeA.on('peer:discovery', (peerInfo) => {
+        expectedPeers.delete(peerInfo.id.toB58String())
+        if (expectedPeers.size === 0) {
+          finish()
+        }
       })
 
-      // connect two dhts
-      const publicPeerId = new PeerId(nodeB._dht.peerInfo.id.id, null, nodeB._dht.peerInfo.id.pubKey)
-      const target = new PeerInfo(publicPeerId)
-      target.multiaddrs = nodeB._dht.peerInfo.multiaddrs
-      nodeA._dht.switch.dial(target, (err) => {
+      // Topology:
+      // A -> B
+      // C -> B
+      nodeA.dial(nodeB.peerInfo, (err) => {
         expect(err).to.not.exist()
       })
+      nodeC.dial(nodeB.peerInfo, (err) => {
+        expect(err).to.not.exist()
+      })
+    })
+  })
+
+  describe('auto dial', () => {
+    setup({
+      connectionManager: {
+        minPeers: 1
+      },
+      config: {
+        peerDiscovery: {
+          autoDial: true,
+          mdns: {
+            enabled: false
+          },
+          webRTCStar: {
+            enabled: false
+          },
+          bootstrap: {
+            enabled: true,
+            list: []
+          }
+        },
+        dht: {
+          enabled: false
+        }
+      }
+    })
+
+    it('should only dial when the peer count is below the low watermark', (done) => {
+      const bootstrap = nodeA._discovery[0]
+      sinon.stub(nodeA._switch.dialer, 'connect').callsFake((peerInfo) => {
+        nodeA._switch.connection.connections[peerInfo.id.toB58String()] = []
+      })
+
+      bootstrap.emit('peer', nodeB.peerInfo)
+      bootstrap.emit('peer', nodeC.peerInfo)
+
+      // Only nodeB should get dialed
+      expect(nodeA._switch.dialer.connect.callCount).to.eql(1)
+      expect(nodeA._switch.dialer.connect.getCall(0).args[0]).to.eql(nodeB.peerInfo)
+      done()
     })
   })
 })