+++ test/transports.browser.js
@@ -4,6 +4,7 @@
 
 const chai = require('chai')
 chai.use(require('dirty-chai'))
+chai.use(require('chai-checkmark'))
 const expect = chai.expect
 const PeerInfo = require('peer-info')
 const PeerId = require('peer-id')
@@ -16,19 +17,19 @@ const wrtcSupport = self.RTCPeerConnection && ('createDataChannel' in self.RTCPe
 const tryEcho = require('./utils/try-echo')
 
 const Node = require('./utils/bundle-browser')
-const jsonPeerId = require('./fixtures/test-peer.json')
+const { getPeerRelay } = require('./utils/constants')
 
 describe('transports', () => {
   describe('websockets', () => {
     let peerB
-    let peerBMultiaddr = '/ip4/127.0.0.1/tcp/9200/ws/p2p/' + jsonPeerId.id
+    let peerBMultiaddr
     let nodeA
 
     before((done) => {
-      PeerId.createFromPrivKey(jsonPeerId.privKey, (err, id) => {
+      getPeerRelay((err, peerInfo) => {
         expect(err).to.not.exist()
-
-        peerB = new PeerInfo(id)
+        peerB = new PeerInfo(peerInfo.id)
+        peerBMultiaddr = `/ip4/127.0.0.1/tcp/9200/ws/p2p/${peerInfo.id.toB58String()}`
         peerB.multiaddrs.add(peerBMultiaddr)
         done()
       })
@@ -413,29 +414,19 @@ describe('transports', () => {
     it('node1 hangUp node2', (done) => {
       node1.hangUp(peer2, (err) => {
         expect(err).to.not.exist()
-        setTimeout(check, 500)
-
-        function check () {
-          const peers = node1.peerBook.getAll()
-          expect(Object.keys(peers)).to.have.length(1)
-          expect(node1._switch.connection.getAll()).to.have.length(0)
-          done()
-        }
+        const peers = node1.peerBook.getAll()
+        expect(Object.keys(peers)).to.have.length(1)
+        expect(node1._switch.connection.getAll()).to.have.length(0)
+        done()
       })
     })
 
     it('create a third node and check that discovery works', function (done) {
       this.timeout(10 * 1000)
-
-      let counter = 0
-
-      function check () {
-        if (++counter === 3) {
-          expect(node1._switch.connection.getAll()).to.have.length(1)
-          expect(node2._switch.connection.getAll()).to.have.length(1)
-          done()
-        }
-      }
+      const expectedPeers = [
+        node1.peerInfo.id.toB58String(),
+        node2.peerInfo.id.toB58String()
+      ]
 
       PeerId.create((err, id3) => {
         expect(err).to.not.exist()
@@ -444,13 +435,18 @@ describe('transports', () => {
         const ma3 = '/ip4/127.0.0.1/tcp/14444/ws/p2p-websocket-star/p2p/' + id3.toB58String()
         peer3.multiaddrs.add(ma3)
 
-        node1.on('peer:discovery', (peerInfo) => node1.dial(peerInfo, check))
-        node2.on('peer:discovery', (peerInfo) => node2.dial(peerInfo, check))
+        // 2 connects and 1 start
+        expect(3).checks(done)
 
         const node3 = new Node({
           peerInfo: peer3
         })
-        node3.start(check)
+        node3.on('peer:connect', (peerInfo) => {
+          expect(expectedPeers).to.include(peerInfo.id.toB58String()).mark()
+        })
+        node3.start((err) => {
+          expect(err).to.not.exist().mark()
+        })
       })
     })
   })