+++ test/utils/bundle-browser.js
@@ -61,6 +61,7 @@ class Node extends libp2p {
       },
       config: {
         peerDiscovery: {
+          autoDial: true,
           webRTCStar: {
             enabled: true
           },